Refactor: 데이터 처리코드  옮기기
