+++ src/components/common/Header/Header.tsx
@@ -3,68 +3,16 @@
 import Image from 'next/image';
 import Link from 'next/link';
 import { useRouter } from 'next/navigation';
-import { useState, useCallback, useEffect } from 'react';
+import { useState, useEffect } from 'react';
 import { SEARCH_ICON, HEADER_LOGO } from '@/utils/constants';
 import Cookies from 'js-cookie';
-import { getUserAlert, putUserAlert } from '@/apis/alert';
 import Notification from './notification';
 import styles from './Header.module.scss';
-import { ButtonListType, NotificationDataType } from './types';
+import { ButtonListType } from './types';
 
 export default function Header() {
   const router = useRouter();
-  /* 
-  notificationListData 샘플 데이터
-  const data = [
-    {
-      id: '1',
-      name: 'test1',
-      startsAt: '2024-04-16T15:00:00.000Z',
-      createdAt: '2024-04-15T15:06:15.633Z',
-      workhour: 3,
-      result: 'accepted',
-      read: false,
-    },
-    {
-      id: '2',
-      name: 'test2',
-      startsAt: '2024-04-16T15:00:00.000Z',
-      createdAt: '2024-04-12T15:06:15.633Z',
-      workhour: 6,
-      result: 'rejected',
-      read: false,
-    },
-    {
-      id: '3',
-      name: 'test3',
-      startsAt: '2024-04-16T15:00:00.000Z',
-      createdAt: '2024-03-19T15:06:15.633Z',
-      workhour: 12,
-      result: 'rejected',
-      read: false,
-    },
-    {
-      id: '4',
-      name: 'test4',
-      startsAt: '2024-04-16T15:00:00.000Z',
-      createdAt: '2023-03-15T15:06:15.633Z',
-      workhour: 4,
-      result: 'rejected',
-      read: false,
-    },
-    {
-      id: '5',
-      name: 'test5',
-      startsAt: '2024-04-16T15:00:00.000Z',
-      createdAt: '2024-04-16T15:06:15.633Z',
-      workhour: 4,
-      result: 'rejected',
-      read: false,
-    },
-  ]; */
   const [memberType, setMemberType] = useState('none');
-  const [notificationData, setNotificationData] = useState<NotificationDataType[]>([]);
-  const [notificationNumber, setNotificationNumber] = useState(notificationData.length);
   const [input, setInput] = useState('');
   const logout = () => {
     const removeData = ['id', 'token', 'type', 'shopId', 'userId'];
@@ -97,29 +45,6 @@ export default function Header() {
     },
   };
   const buttonType = BUTTON_LIST[memberType];
-  const handleClickNotification = useCallback(async (id: string, isRead: boolean) => {
-    if (isRead) {
-      return;
-    }
-    /* api function */
-    const status = await putUserAlert(id);
-    if (status === '400' || status === '404') {
-      throw new Error(`잘못된 접근: ${status}`);
-    }
-    if (status === '403') {
-      alert('로그인 해주세요.');
-      router.push('/signin');
-      return;
-    }
-    setNotificationData((previousData) =>
-      previousData.map((notification) => (notification.id === id ? { ...notification, read: true } : notification)),
-    );
-    setNotificationNumber((previousCount) => previousCount - 1);
-  }, []);
-
-  const changeNotificationData = useCallback(() => {
-    setNotificationData((previousData) => previousData.filter((notification) => !notification.read && notification));
-  }, []);
   const handleInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
     setInput(event.target.value);
   };
@@ -135,21 +60,6 @@ export default function Header() {
       return;
     }
     setMemberType('none');
-  }, []);
-
-  useEffect(() => {
-    const getData = async () => {
-      const token = Cookies.get('token');
-      if (token) {
-        const alertData = await getUserAlert();
-        if (typeof alertData === 'number') {
-          throw new Error(`Alert API Error ${alertData}`);
-        }
-        setNotificationData(alertData);
-        setNotificationNumber(alertData.length);
-      }
-    };
-    getData();
   }, []);
 
   return (
@@ -180,14 +90,7 @@ export default function Header() {
               </Link>
             ),
           )}
-          {buttonType.notification && (
-            <Notification
-              notificationData={notificationData}
-              notificationNumber={notificationNumber}
-              handleClickNotification={handleClickNotification}
-              changeNotificationData={changeNotificationData}
-            />
-          )}
+          {buttonType.notification && <Notification />}
         </div>
       </div>
     </div>
